# services/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Route inventory with explicit operations schemas for the admin client. Unspecified payloads remain unknown.",
    "title": "TripRaft API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/expenses/group/{group_id}": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "expenses_sql_get_group_expenses_get_api_v1_expenses_group_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/expenses/groups/{group_id}": {
      "delete": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_delete_group_delete_api_v1_expenses_groups_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      },
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_get_group_get_api_v1_expenses_groups_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      },
      "put": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_update_group_put_api_v1_expenses_groups_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
    "/api/v1/expenses/groups/{group_id}/full": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_get_group_full_get_api_v1_expenses_groups_group_id_full",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
    "/api/v1/expenses/groups/{group_id}/leave": {
      "post": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_leave_group_post_api_v1_expenses_groups_group_id_leave",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
    "/api/v1/expenses/groups/{group_id}/members": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_get_group_members_get_api_v1_expenses_groups_group_id_members",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      },
      "post": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_add_member_post_api_v1_expenses_groups_group_id_members",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
    "/api/v1/expenses/groups/{group_id}/members/{member_id}": {
      "delete": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "groups_sql_remove_member_delete_api_v1_expenses_groups_group_id_members_member_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/expenses/invitations/groups/{group_id}": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "invitations_sql_get_group_invitations_get_api_v1_expenses_invitations_groups_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/expenses/settlements/group/{group_id}": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "settlements_sql_get_group_settlements_get_api_v1_expenses_settlements_group_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
    "/api/v1/expenses/settlements/group/{group_id}/balances": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "settlements_sql_get_group_balances_get_api_v1_expenses_settlements_group_group_id_balances",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/expenses/{expense_id}": {
      "delete": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "expenses_sql_delete_expense_delete_api_v1_expenses_expense_id",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      },
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "expenses_sql_get_expense_get_api_v1_expenses_expense_id",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      },
      "put": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "expenses_sql_update_expense_put_api_v1_expenses_expense_id",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/group-planner/groups/{group_id}": {
      "delete": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_groups_delete_group_delete_api_v1_group_planner_groups_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      },
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_groups_get_group_get_api_v1_group_planner_groups_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      },
      "put": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_groups_update_group_put_api_v1_group_planner_groups_group_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/group-planner/groups/{group_id}/invitations": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_invitations_get_group_invitations_get_api_v1_group_planner_groups_group_id_invitations",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/group-planner/groups/{group_id}/leave": {
      "post": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_groups_leave_group_post_api_v1_group_planner_groups_group_id_leave",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
    "/api/v1/group-planner/groups/{group_id}/members": {
      "get": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_groups_get_group_members_get_api_v1_group_planner_groups_group_id_members",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
    "/api/v1/group-planner/groups/{group_id}/members/{member_id}": {
      "delete": {
        "description": "Registered route. Response payload is unspecified unless an explicit schema is attached.",
        "operationId": "gp_groups_remove_member_delete_api_v1_group_planner_groups_group_id_members_member_id",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Response"
          }
        }
      }
    },
...
  }
}